cm-render: Layer 3↔2 bridge — fix WidgetDescriptor 5-field mislabel + production to_render_widget
The 6 style fields the layer-2 render code (`FUN_005d7aa0` / `packed_widget::Widget`)
consumes were mislabelled on `WidgetDescriptor`, and `spawn_widget` silently dropped
them on the floor — every screen builder's style output was lost between spawn and
render. Verified against the GDI-runtime populator sub_005d76c0 (ret 0x54 = 21 stack
args, matching the cm0102 decomp's 22 params); the field-offset writes at
005d78ad..005d791e pin every widget slot unambiguously.

Field renames (was → is, per FUN_005d76c0's `param_1[N] = param_J` writes):

| caller arg | offset | old name                | new name         |
|-----------:|--------|-------------------------|------------------|
|          1 | +0x0c  | `kind: u16`             | `kind: u32`      |
|          8 | +0x38  | `flags: u32`            | `style_byte: u32`|
|          9 | +0x72  | `unk9: i32`             | `colour_a: u16`  |
|         10 | +0x74  | `unk10: i32`            | `colour_b: u16`  |
|         11 | +0x3c  | `font_id: u8`           | `text_style: u32`|
|         12 | +0x76  | `enabled: bool`         | `label_ink: u16` |
|         13 | +0x78  | `fg_color: u32`         | `pattern: u16`   |
|         15 | +0x40  | `extra: Vec<u8>`        | `slot_40: i32`   |

The 3 correctly-placed fields are preserved as-is:
* `text` (arg14, strcpy → +0x80)
* `msg_id` (arg16, +0x08 — dispatcher.rs reads this as the click cmd)
* `userdata_id` (arg17, +0x48 — dispatcher.rs reads this as the payload)

Struct merge decision: `widget_pool::Widget` (spawn-time state, layout arrays,
z-order children) and `packed_widget::Widget` (render-time state, saved_bg,
cached_text, alt_hover) hold DISJOINT runtime fields. Kept separate; added the
6 style fields + `label` to `widget_pool::Widget` so the bridge stays 1:1 with
no defaults or reinterpretation. `pool_to_render::to_render_widget` is the
single canonical bridge — see its module doc for the rationale.

Fixes in this commit:

* `spawn_widget` (`FUN_00549580`) now copies every populated field onto the
  resulting `Widget`, matching FUN_005d76c0's per-offset writes. Previously
  only rect + parent_area + cached_grid_index were kept; the 6 style fields
  reached the render layer as zeros.

* Swap-gate reads the correct arg (arg8 = `style_byte` at +0x38, previously
  read arg8 via the misnamed `flags` — same slot, name now honest).

* Production `pool_to_render::to_render_widget` bridge (new module).
  Roundtrips every field. Test proves every style field carries through.

* Callers updated per each site's own asm arg order:
    - `screen_nav_back_next.rs` (FUN_005d75b0 nav-bar): every arg annotated;
      label_ink now 3 (per C literal), pattern now DAT_00acdf74 (gold).
    - `screen_news.rs` spawn_label/spawn_button (sub_00770170): every arg
      annotated; label_ink now the captured `col=NNNNN`; unobserved slots
      default to 0 pending a capture that surfaces them.
    - `screen_menu_bar.rs` MBR holder + sidebar buttons + dropdown items
      (FUN_00745540): every FUN_00549580 call annotated per its asm site.

* Test-only `to_render_widget` stubs in the 3 screen files removed;
  callers now import `crate::pool_to_render::to_render_widget`.

* Downstream code updated for renamed fields:
    - `view_render.rs`: `d.flags` → `d.kind` (widget flags dword),
      `d.font_id` → `d.text_style` (+ disabled bit 0x20 for the tflags
      12→44 distinction), `d.enabled` → text_style bit, `d.fg_color` →
      `d.label_ink`. `w.flags` (rflags dword on the Widget) preserved for
      the highlight bit tests. `tab_button` sets highlight via
      `Widget.flags`, not `WidgetDescriptor.kind`.
    - `gen_screen_types.rs`: kind carries OR-ed rflags from captures.
    - `bin/dump_screen_geometry.rs`, `bin/render_news.rs`,
      `tests/news_pixel_diff_against_exe.rs`, `cm-ui-app/src/screens.rs`:
      renamed field reads.

Test count: 447 cm-render lib tests passing (previously 447 — same count,
1 pre-existing packed_glyph failure unrelated to this commit). New tests:
2 in `pool_to_render::tests` (every-field-roundtrip and
spawn→bridge→render). Existing `integration_menu_bar_then_dispatch_news_produces_both_layers`
still green end-to-end.

Cross-referenced with GDI-runtime asm sub_005d76c0 (the populator): the
field-offset writes at 005d78ad..005d791e confirm the caller-arg → widget-offset
mapping. cm0102 decomp `005d7bd0.c` and GDI asm agree on every offset.

+++ crates/cm-ui-app/src/screens.rs
@@ -2080,7 +2080,7 @@ fn blank_with_warning(s: &mut Surface, msg: &str) {
 
 /// Colour palette per widget kind — chosen so the four common kinds are
 /// visually distinct at a glance.
-fn kind_colour(kind: u16) -> (u8, u8, u8) {
+fn kind_colour(kind: u32) -> (u8, u8, u8) {
     use cm_render::widget_pool::*;
     match kind {
         KIND_HEADER => (60, 90, 180),      // navy
@@ -2091,7 +2091,7 @@ fn kind_colour(kind: u16) -> (u8, u8, u8) {
     }
 }
 
-fn kind_name(kind: u16) -> &'static str {
+fn kind_name(kind: u32) -> &'static str {
     use cm_render::widget_pool::*;
     match kind {
         KIND_HEADER => "HDR",
